+++ docs/source/changelog.md
@@ -2,6 +2,7 @@
 
 ## Latest
 - Switch back to using numpy nullable dtypes for Pandas as default.
+- Ensure that indices get created in same schema as corresponding table (IBM Db2).
 - Fix private method `SQLTableStore.get_stage_hash`.
 
 ## 0.6.1 (2023-07-19)